Moved heart_rate_imposer to server and added heart_rate_deamon that will be a deamon that loads the models and responds to requests from server through a message queue.
