+++ app.py
@@ -1,29 +1,49 @@
-from flask import Flask, render_template, Response
-import sys
-import random
-from server.heart_rate_imposer import HeartRateImposer
-from keras.models import model_from_json
+from flask import Flask, request, Response, abort, send_file, jsonify
+import os, subprocess, re, sys
 
 app = Flask(__name__)
 
 video_file = sys.argv[1]
 opencv_path = sys.argv[2]
 
-def gen_http_frame(gen_frames):
-    for frame in gen_frames():
-        yield (b'--frame\r\n'
-               b'Content-Type: image/jpeg\r\n\r\n' + frame + b'\r\n')
-
 @app.route('/')
 def index():
-    return "Hello World"
+    return 'Hello World'
 
 @app.route('/video')
-def stream():
-    hri = HeartRateImposer(video_file, opencv_path)
-    return Response(gen_http_frame(hri.gen_heartrate_frames),
-                    mimetype='multipart/x-mixed-replace; boundary=frame')
+def video():
+    FNULL = open(os.devnull, 'w')
 
+    hri_cmd = ['python', 'heart_rate_deamon.py', video_file, opencv_path]
+    opencv_ps = subprocess.Popen(hri_cmd, stdout=subprocess.PIPE, stderr=FNULL)
+
+    line = opencv_ps.stdout.readline()
+    [width, height, fps] = line.split(' ')
+
+    cmd = 'ffmpeg -f rawvideo -pix_fmt bgr24 -s WIDTHxHEIGHT -r 30 -i - -f ogg -an -qscale:v 10 pipe:1'
+    cmd = cmd.replace('WIDTH', width).replace('HEIGHT', height)
+    cmd = cmd.split(' ')
+    
+    # lets you skip forward
+    start = request.args.get("start") or 0
+    def generate():
+        proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stdin=opencv_ps.stdout)
+        try:
+            f = proc.stdout
+            byte = f.read(512)
+            while byte:
+                yield byte
+                byte = f.read(512)
+        finally:
+            proc.kill()
+
+    return Response(response=generate(),
+                    status=200,
+                    mimetype='video/ogg',
+                    headers={'Access-Control-Allow-Origin': '*',
+                             "Content-Type": "video/ogg",
+                             "Content-Disposition": "inline",
+                             "Content-Transfer-Enconding": "binary"})
 
 if __name__ == '__main__':
-    app.run(host='0.0.0.0', debug=True)
+    app.run(host='0.0.0.0', threaded=True, debug=True)
