+++ server/heart_rate_imposer.py
@@ -108,56 +108,3 @@ class HeartRateImposer(object):
     self._add_face_detection(to_file)
     self._copy_audio(self.from_file, to_file)
 
-  def _add_face_detection(self, to_file):
-    face_cascade = cv2.CascadeClassifier(self.opencv_path +
-                     'data/haarcascades/haarcascade_frontalface_default.xml')
-
-    cap = cv2.VideoCapture(self.from_file)
-
-    w = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
-    h = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
-    fourcc = int(cap.get(cv2.CAP_PROP_FOURCC))
-    fps = cap.get(cv2.CAP_PROP_FPS)
-
-    writer = cv2.VideoWriter(to_file, fourcc, fps, (w, h))
-
-    for heartrate in self.heartrates:
-      ret, frame = cap.read()
-
-      if not ret or frame is None:
-        break
-
-      gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-      faces = face_cascade.detectMultiScale(gray, 1.3, 5)
-
-      for (x,y,w,h) in faces:
-        cv2.rectangle(frame, (x,y), (x+w,y+h), (0,0,255), 2)
-        cv2.putText(frame, str(int(heartrate)) + ' bpm', (x+(w*1)/4, y+h+20),
-                    cv2.FONT_HERSHEY_SIMPLEX, 0.75, (0, 0, 255), 2)
-
-      writer.write(frame)
-
-    cap.release()
-    cv2.destroyAllWindows()
-
-  def _copy_audio(self, from_file, to_file):
-    temp_audio_file = 'audio_' + to_file
-    subprocess.call(['ffmpeg', '-i', to_file, '-i', from_file, '-c', 'copy',
-                     '-map', '0:0', '-map', '1:1', '-shortest', temp_audio_file])
-    subprocess.call(['mv', temp_audio_file, to_file])
-
-if __name__ == "__main__":
-    # python heart_rate_imposer video_file opencv_path
-    video_file = sys.argv[1]
-    opencv_path = sys.argv[2]
-
-    hri = HeartRateImposer(video_file, opencv_path)
-    hri.pipe_heartrate_frames()
-
-
-
-
-
-
-
-
